# minimize allocations in evalh routines