@ahbarnett 2023-01-20
Ok, now I don't understand your claim of resounding success. You have made `realadap_lxvm` about 10x faster than `realadap`, so now it's about the same speed as `imshcorr`, right, at M=100. I guess at M=10 it's still 5x faster. But I have a secret weapon about to be git pushed... stay tuned.