@ahbarnett 2023-01-20
In commit 394e491 you'll see I switched to a faster rootfinder, recovering about 15x speedup over `realadap_lxvm` even up to M=100:

I still don't really see why your `fourier_kernel` is slower in the 1000-targ test but so much faster in quadgk...

```
julia> include("bench_Cont1DBZ.jl")

bench Cont1DBZ with M=1.
	Eval at 1000 targs...
evalh_ref:	  48.982 μs (14 allocations: 212.69 KiB)
evalh_wind:	  23.544 μs (3 allocations: 47.25 KiB)
evalh:    	  21.070 μs (7 allocations: 63.38 KiB)
evalh 0.141576 G mode-targs/sec
fourier_kernel:    	  13.245 μs (1 allocation: 15.75 KiB)
	Quadr meths:
realadap ω=0.5 η=1e-06 tol=1e-08:    672.164 μs (21008 allocations: 1.33 MiB)
	Quadr meths:
realadap_lxvm ω=0.5 η=1e-06 tol=1e-08:    72.756 μs (8 allocations: 9.16 KiB)
roots (matrix size 2M+1=3):    1.689 μs (21 allocations: 2.66 KiB)
roots_best (matrix size 2M+1=3):    169.869 ns (4 allocations: 432 bytes)
imshcorr same ω and η as above, NPTR=30:     5.238 μs (34 allocations: 5.52 KiB)

bench Cont1DBZ with M=10.
	Eval at 1000 targs...
evalh_ref:	  437.943 μs (86 allocations: 1.32 MiB)
evalh_wind:	  52.319 μs (3 allocations: 47.25 KiB)
evalh:    	  29.015 μs (7 allocations: 63.38 KiB)
evalh 0.719548 G mode-targs/sec
fourier_kernel:    	  33.062 μs (1 allocation: 15.75 KiB)
	Quadr meths:
realadap ω=0.5 η=1e-06 tol=1e-08:    6.533 ms (112449 allocations: 7.11 MiB)
	Quadr meths:
realadap_lxvm ω=0.5 η=1e-06 tol=1e-08:    636.457 μs (9 allocations: 36.84 KiB)
roots (matrix size 2M+1=21):    132.129 μs (21 allocations: 88.36 KiB)
roots_best (matrix size 2M+1=21):    19.306 μs (4 allocations: 1.56 KiB)
imshcorr same ω and η as above, NPTR=30:     52.909 μs (115 allocations: 44.58 KiB)

bench Cont1DBZ with M=100.
	Eval at 1000 targs...
evalh_ref:	  4.284 ms (806 allocations: 12.39 MiB)
evalh_wind:	  326.765 μs (3 allocations: 47.25 KiB)
evalh:    	  99.808 μs (7 allocations: 63.38 KiB)
evalh 1.98772 G mode-targs/sec
fourier_kernel:    	  242.356 μs (1 allocation: 15.75 KiB)
	Quadr meths:
realadap ω=0.5 η=1e-06 tol=1e-08:    633.696 ms (2005332 allocations: 126.81 MiB)
	Quadr meths:
realadap_lxvm ω=0.5 η=1e-06 tol=1e-08:    56.750 ms (12 allocations: 607.97 KiB)
roots (matrix size 2M+1=201):    39.996 ms (24 allocations: 1.95 MiB)
roots_best (matrix size 2M+1=201):    1.931 ms (4 allocations: 13.19 KiB)
imshcorr same ω and η as above, NPTR=30:     3.780 ms (847 allocations: 2.59 MiB)
```

My `roots_best` switches to a different method, not quite as fast, for M>110. Eg, check out:

```
bench Cont1DBZ with M=200.
	Eval at 1000 targs...
evalh_ref:	  8.512 ms (1606 allocations: 24.69 MiB)
evalh_wind:	  628.973 μs (3 allocations: 47.25 KiB)
evalh:    	  178.907 μs (7 allocations: 63.38 KiB)
evalh 2.2572 G mode-targs/sec
fourier_kernel:    	  472.669 μs (1 allocation: 15.75 KiB)
	Quadr meths:
realadap ω=0.5 η=1e-06 tol=1e-08:    48.339 s (80000057 allocations: 4.92 GiB)
	Quadr meths:
realadap_lxvm ω=0.5 η=1e-06 tol=1e-08:    531.294 ms (14 allocations: 3.97 MiB)
roots (matrix size 2M+1=401):    227.145 ms (24 allocations: 7.55 MiB)
roots_best (matrix size 2M+1=401):    37.298 ms (14 allocations: 76.97 KiB)
imshcorr same ω and η as above, NPTR=30:     42.938 ms (1835 allocations: 10.20 MiB)
```

quadgk borks on my evalh but not your fourier_kernel, weirdly.

Anyway, on to n>1 matrix NEVP cases...